更新 music.py

+++ music.py
@@ -66,7 +66,7 @@ for i, lyric in enumerate(input_lyrics.replace('\n', '')):
 while len(lyrics[-1]) % batch_size != 0:
     lyrics[-1].append(ord('#'))
 print(lyrics)
-lyrics = paddle.to_tensor(lyrics)
+lyrics = paddle.to_tensor(lyrics,dtype="int32")
 print(lyrics)
 params_dict = paddle.load('Midi_Model/best_model')
 midi_model.set_dict(params_dict)
